+++ packages/dom-snapshot/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+- fix missing resources when dealing with fake shadow dom
 
 ## 3.4.3 - 2020/4/5